requirements.txt
attrs==20.3.0
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
